+++ sample/getObjectUrl.php
@@ -18,7 +18,7 @@ $cosClient = new Qcloud\Cos\Client(
 $local_path = "/data/exampleobject";
 
 try {
-    $bucket = "examplebucket-1250000000"; //存储桶，格式：BucketName-APPID
+    $bucket = "examplebucket-1250000000"; //存储桶，存储桶名称，由BucketName-Appid 组成，可以在COS控制台查看 https://console.cloud.tencent.com/cos5/bucket
     $key = "exampleobject"; //对象在存储桶中的位置，即对象键
     $signedUrl = $cosClient->getObjectUrl(
         $bucket,