+++ sample/imageMogr.php
@@ -17,7 +17,7 @@ try {
     $imageMogrTemplate->thumbnailByScale(50);
     $imageMogrTemplate->rotate(50);
     $result = $cosClient->getObject(array(
-        'Bucket' => 'examplebucket-125000000', //格式：BucketName-APPID
+        'Bucket' => 'examplebucket-125000000', //存储桶名称，由BucketName-Appid 组成，可以在COS控制台查看 https://console.cloud.tencent.com/cos5/bucket
         'Key' => 'exampleobject',
         'ImageHandleParam' => $imageMogrTemplate->queryString(),
         'SaveAs' => '/data/exampleobject',